+++ pkg/cmd/release/shared/options_test.go
@@ -1,4 +1,4 @@
-package attestation
+package shared
 
 import (
 	"errors"
