+++ pkg/cmd/release/shared/policy.go
@@ -1,4 +1,4 @@
-package attestation
+package shared
 
 import (
 	"fmt"
