+++ pkg/cmd/release/verify-asset/verify-asset.go
@@ -14,15 +14,14 @@ import (
 	"github.com/cli/cli/v2/pkg/cmd/attestation/artifact"
 	att_io "github.com/cli/cli/v2/pkg/cmd/attestation/io"
 	"github.com/cli/cli/v2/pkg/cmd/attestation/verification"
-	"github.com/cli/cli/v2/pkg/cmd/release/attestation"
 	"github.com/cli/cli/v2/pkg/cmd/release/shared"
 
 	"github.com/cli/cli/v2/pkg/cmdutil"
 	"github.com/spf13/cobra"
 )
 
-func NewCmdVerifyAsset(f *cmdutil.Factory, runF func(*attestation.AttestOptions) error) *cobra.Command {
-	opts := &attestation.AttestOptions{}
+func NewCmdVerifyAsset(f *cmdutil.Factory, runF func(*shared.AttestOptions) error) *cobra.Command {
+	opts := &shared.AttestOptions{}
 
 	cmd := &cobra.Command{
 		Use:    "verify-asset <tag> <file-path>",
@@ -56,14 +55,14 @@ func NewCmdVerifyAsset(f *cmdutil.Factory, runF func(*attestation.AttestOptions)
 				return err
 			}
 
-			*opts = attestation.AttestOptions{
+			*opts = shared.AttestOptions{
 				TagName:       opts.TagName,
 				AssetFilePath: opts.AssetFilePath,
 				Repo:          baseRepo.RepoOwner() + "/" + baseRepo.RepoName(),
 				APIClient:     api.NewLiveClient(httpClient, hostname, logger),
 				Limit:         10,
 				Owner:         baseRepo.RepoOwner(),
-				PredicateType: attestation.ReleasePredicateType,
+				PredicateType: shared.ReleasePredicateType,
 				Logger:        logger,
 				HttpClient:    httpClient,
 				BaseRepo:      baseRepo,
@@ -86,7 +85,7 @@ func NewCmdVerifyAsset(f *cmdutil.Factory, runF func(*attestation.AttestOptions)
 
 			opts.TrustedRoot = td
 
-			ec, err := attestation.NewEnforcementCriteria(opts)
+			ec, err := shared.NewEnforcementCriteria(opts)
 			if err != nil {
 				opts.Logger.Println(opts.Logger.ColorScheme.Red("✗ Failed to build policy information"))
 				return err
@@ -109,7 +108,7 @@ func NewCmdVerifyAsset(f *cmdutil.Factory, runF func(*attestation.AttestOptions)
 	return cmd
 }
 
-func verifyAssetRun(opts *attestation.AttestOptions) error {
+func verifyAssetRun(opts *shared.AttestOptions) error {
 	ctx := context.Background()
 
 	if opts.SigstoreVerifier == nil {
@@ -156,7 +155,7 @@ func verifyAssetRun(opts *attestation.AttestOptions) error {
 	opts.Logger.Printf("Resolved %s to %s\n", opts.TagName, releaseRefDigest.DigestWithAlg())
 
 	// Attestation fetching
-	attestations, logMsg, err := attestation.GetAttestations(opts, releaseRefDigest.DigestWithAlg())
+	attestations, logMsg, err := shared.GetAttestations(opts, releaseRefDigest.DigestWithAlg())
 	if err != nil {
 		if errors.Is(err, api.ErrNoAttestationsFound) {
 			opts.Logger.Printf(opts.Logger.ColorScheme.Red("✗ No attestations found for subject %s\n"), releaseRefDigest.DigestWithAlg())
@@ -167,13 +166,13 @@ func verifyAssetRun(opts *attestation.AttestOptions) error {
 	}
 
 	// Filter attestations by tag
-	filteredAttestations, err := attestation.FilterAttestationsByTag(attestations, opts.TagName)
+	filteredAttestations, err := shared.FilterAttestationsByTag(attestations, opts.TagName)
 	if err != nil {
 		opts.Logger.Println(opts.Logger.ColorScheme.Red(err.Error()))
 		return err
 	}
 
-	filteredAttestations, err = attestation.FilterAttestationsByFileDigest(filteredAttestations, opts.Repo, opts.TagName, fileDigest.Digest())
+	filteredAttestations, err = shared.FilterAttestationsByFileDigest(filteredAttestations, opts.Repo, opts.TagName, fileDigest.Digest())
 	if err != nil {
 		opts.Logger.Println(opts.Logger.ColorScheme.Red(err.Error()))
 		return err
@@ -187,7 +186,7 @@ func verifyAssetRun(opts *attestation.AttestOptions) error {
 	opts.Logger.Printf("Loaded %s from GitHub API\n", text.Pluralize(len(filteredAttestations), "attestation"))
 
 	// Verify attestations
-	verified, errMsg, err := attestation.VerifyAttestations(*releaseRefDigest, filteredAttestations, opts.SigstoreVerifier, opts.EC)
+	verified, errMsg, err := shared.VerifyAttestations(*releaseRefDigest, filteredAttestations, opts.SigstoreVerifier, opts.EC)
 
 	if err != nil {
 		opts.Logger.Println(opts.Logger.ColorScheme.Red(errMsg))
